Create fangraphs.depth_charts module to separate from fangraphs.teams
