+++ docs/fangraphs.depth_charts.rst
@@ -1,0 +1,7 @@
+Fangraphs.depth_charts Module
+=============================
+
+.. automodule:: fangraphs.depth_charts
+    :members:
+    :undoc-members:
+    :show-inheritance:
